+++ lib/src/core/digit.dart
@@ -1,7 +1,7 @@
 /// Digit - any one of the ten numbers 0 to 9.
 ///
-/// For example, the input values "1" and "123" are valid; whereas the values
-/// "A", "12A", "123.34", "9.99" are invalid.
+/// For example, the values "1" and "123" are valid; whereas the values "A",
+/// "12A", "123.34", "9.99" are invalid.
 ///
 /// Blank field - null value - is a valid input.
 ///
@@ -9,8 +9,9 @@
 class Digit {
   /// Validates an optional digit-only field.
   ///
-  /// [non] is the error message in case of a non-digit value.
-  const Digit({String non = 'digit-only field'}) : _nonDigit = non;
+  /// [non] is the error message in case a non-digit character is found in the
+  /// input data; defaults to 'non-digit character'.
+  const Digit({String? non}) : _nonDigit = non ?? 'non-digit character';
 
   final String _nonDigit;