+++ lib/src/core/req.dart
@@ -1,21 +1,22 @@
-/// Mandatory field.
+/// Mandatory input data.
 ///
-/// **Note:** both the **null** value and empty are **invalid inputs**.
+/// It will reject null or empty input data.
 class Req {
   /// Non-blank field.
   ///
-  /// [blank] is the error message in case of a null or empty input.
-  const Req({String blank = 'required'}) : _blank = blank;
+  /// [blank] the error message in case of null or empty input data; defaults to
+  /// 'required'.
+  const Req({String? blank}) : _blank = blank ?? 'required';
 
   final String _blank;
 
-  /// Valid - returns null - if [value] is neither null nor empty.
-  String? call(Object? value) =>
-      (value == null || _isEmpty(value)) ? _blank : null;
+  /// Valid - returns null - if [input] is neither null nor empty.
+  String? call(Object? input) =>
+      (input == null || _isEmpty(input)) ? _blank : null;
 
-  bool _isEmpty(Object value) {
-    return (value is String && value.isEmpty) ||
-        (value is Iterable && value.isEmpty) ||
-        (value is Map && value.isEmpty);
+  bool _isEmpty(Object input) {
+    return (input is String && input.isEmpty) ||
+        (input is Iterable && input.isEmpty) ||
+        (input is Map && input.isEmpty);
   }
 }