+++ lib/src/core/trim.dart
@@ -4,10 +4,10 @@ import 'package:formdator/formdator.dart';
 ///
 /// It does not validate anything! It just trims input data and forwards it.
 class Trim {
-  /// Will remove any leading or trailing whitspace from input before forwarding
-  /// it to [val].
+  /// Will remove any leading or trailing whitespaces from input before
+  /// forwarding it to [val].
   ///
-  /// [val] a validator for string input data.
+  /// [val] a validator for [String] input data.
   const Trim(ValStr val) : _val = val;
 
   final ValStr _val;