+++ lib/src/net/req_email.dart
@@ -7,14 +7,16 @@ import 'email.dart';
 class ReqEmail {
   /// Ensures a non-blank and well-formed email value.
   ///
-  /// [blank] is the blank field message; [mal], the malformed email message.
-  ReqEmail({String blank = 'required', String mal = 'malformed'})
-      : _reqEmail = Rules([Req(blank: blank), Email(mal: mal)]);
+  /// [blank] the blank field message; defaults to 'required email'.
+  /// [mal] the malformed email message.
+  ReqEmail({String? blank, String? mal})
+      : _reqEmail =
+            Rules([Req(blank: blank ?? 'required email'), Email(mal: mal)]);
 
   final Rules<String> _reqEmail;
 
-  /// Valid — returns null — if [email] is a non-blank and well-formed email;
-  /// otherwise, returns the blank error message if [email] is null, or the
-  /// malformed error message if [email] is malformed.
-  String? call(String? email) => _reqEmail(email);
+  /// Valid — returns null — if [input] is a non-blank well-formed email;
+  /// otherwise, returns the blank error message if [input] is null, or the
+  /// malformed error message if [input] is malformed.
+  String? call(String? input) => _reqEmail(input);
 }